refactor: move cmd lib functions to lib.go
