+++ cmd/config-cluster-set.go
@@ -64,7 +64,7 @@ with a different base URL will change the base URL for the 'foobar' cluster.`,
 		}
 
 		// Ask user to create file if it does not exist
-		if err := AskToCreate(fileToModify); err != nil {
+		if err := askToCreate(fileToModify); err != nil {
 			if errors.Is(err, UserDeclinedError) {
 				log.Logger.Info().Msgf("user declined creating config file %s, exiting")
 				os.Exit(0)
